+++ src/plugins/alsa/qalsaaudiooutput.cpp
@@ -782,7 +782,7 @@ qint64 QAlsaAudioOutput::elapsedUSecs() const
     if (deviceState == QAudio::StoppedState)
         return 0;
 
-    return clockStamp.elapsed()*1000;
+    return clockStamp.elapsed() * qint64(1000);
 }
 
 void QAlsaAudioOutput::reset()
